Resolved the strict mode If strict mode is not enabled
I was facing the default value  '0000-00-00 00:00:00' of updated_on field. After deep debugging I found, strict mode is enabled.

If strict mode is enabled, '0000-00-00' is not permitted and inserts produce an error, unless IGNORE is given as well. For INSERT IGNORE and UPDATE IGNORE, '0000-00-00' is permitted and inserts produce a warning.